+++ src/main/java/meowmeow/commands/ExitCommand.java
@@ -1,18 +1,18 @@
-package duke.commands;
-import duke.Storage;
-import duke.TaskList;
-import duke.Ui;
+package meowmeow.commands;
+import meowmeow.Storage;
+import meowmeow.TaskList;
+import meowmeow.Ui;
 
 /**
- * <p>Class ListCommand is a concrete class that implements the Command interface.</p>
- * <p>This class is used when the user enters the "list" command.</p>
+ * <p>Class ExitCommand is a concrete class that implements the Command interface.</p>
+ * <p>This class is used when the user enters the bye command.</p>
  * <p>This class is a concrete class because it has an implementation.</p>
  */
-public class ListCommand extends Command {
+public class ExitCommand extends Command {
 
     /**
-     * Method that executes the ListCommand.
-     * Prints out the tasks in the task list to the user interface.
+     * Method that executes the ExitCommand.
+     * Prints a goodbye message to the user interface.
      *
      * @param tasks the task list to be operated on by the command.
      *              The task list is used to delete the task from the task list.
@@ -23,15 +23,15 @@ public class ListCommand extends Command {
      */
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
-        tasks.printTaskList();
+        System.out.println("UwU Byebyeeee! Come back soon... Meowmeow misses you already =^._.^= ");
     }
 
     /**
-     * isExit method that returns false for ListCommand.
-     * @return false for ListCommand.
+     * isExit method that returns true for ExitCommand.
+     * @return true for ExitCommand.
      */
     @Override
     public boolean isExit() {
-        return false;
+        return true;
     }
 }
